Updated Booking and Search Pages
Added a little card thing at the top of the search page to visually show how the data, time, and party size of the saved reservation data. It automatically updates whenever a change in reservation data is made.

Dialogue was added to both cancel and edit buttons in the Booking Page.

Also added 'loadData()' to the save reservation thingy in main.dart so that the little card thing in the search page shows accurate data.

Last and most importantly:
giordanachos.jpg was fixed (it wasnt in the pubspec.yaml)