literally jake's branch + my own work
There was a weird problem with my gradle or something. An error popping up and the only way to fix it is to create a new project. Kinda said fuck it and just copied Jake's branch (because his can connect to the database) and then swapped some files with my own (search page, menu page, reserve page).

The reserve page now asks the user for a name before they are able to reserve a specific table.